+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Unreleased
 
-- Fix for #8 - Coercion is done only if `data` is not null or not undefined.
+- Fix for #8 - Coercion is done only, if `data` is not null or not undefined.